Rename --net and --network to --tcp
