+++ bin/loadtest.js
@@ -36,7 +36,7 @@ const options = stdio.getopt({
 	cert: {args: 1, description: 'The client certificate to use'},
 	quiet: {description: 'Do not log any messages'},
 	cores: {args: 1, description: 'Number of cores to use', default: getHalfCores()},
-	net: {description: 'Use network sockets (experimental)'},
+	tcp: {description: 'Use TCP sockets (experimental)'},
 	agent: {description: 'Use a keep-alive http agent (deprecated)'},
 	debug: {description: 'Show debug messages (deprecated)'},
 });
