+++ lib/loadtest.js
@@ -29,6 +29,7 @@ const SHOW_INTERVAL_MS = 5000;
  *	 - debug: show debug messages (deprecated).
  *	 - requestGenerator: use a custom function to generate requests.
  *	 - statusCallback: function called after every request.
+ *	 - tcp: use TCP sockets (experimental).
  * - `callback`: optional `function(result, error)` called if/when the test finishes;
  * if not present a promise is returned.
  */
