+++ lib/options.js
@@ -46,7 +46,7 @@ class Options {
 		this.proxy = options.proxy || configuration.proxy
 		this.quiet = options.quiet || configuration.quiet
 		this.statusCallback = options.statusCallback
-		this.net = options.net || configuration.net
+		this.tcp = options.tcp || configuration.tcp
 		if (!this.maxRequests && !this.maxSeconds) {
 			this.maxSeconds = 10
 		}
