+++ lib/pool.js
@@ -1,5 +1,5 @@
 import {HttpClient} from './httpClient.js'
-import {NetworkClient} from './networkClient.js'
+import {TcpClient} from './tcpClient.js'
 import {WebsocketClient} from './websocket.js'
 import {HighResolutionTimer} from './hrtimer.js'
 
@@ -50,8 +50,8 @@ export class Pool {
 		if (this.options.url.startsWith('ws:')) {
 			return new WebsocketClient(this.loadTest)
 		}
-		if (this.options.net) {
-			return new NetworkClient(this.loadTest)
+		if (this.options.tcp) {
+			return new TcpClient(this.loadTest)
 		}
 		return new HttpClient(this.loadTest);
 	}
