+++ lib/result.js
@@ -34,8 +34,8 @@ export class Result {
 		this.maxSeconds = parseInt(options.maxSeconds)
 		this.concurrency = parseInt(options.concurrency)
 		this.clients = latency.clients
-		if (options.net) {
-			this.agent = 'sockets'
+		if (options.tcp) {
+			this.agent = 'tcp'
 		} else if (options.agentKeepAlive) {
 			this.agent = 'keepalive'
 		} else {
