+++ lib/tcpClient.js
@@ -10,13 +10,13 @@ const forbiddenOptions = [
 
 
 /**
- * A client for an HTTP connection, using raw sockets. Constructor:
+ * A client for an HTTP connection, using TCP sockets. Constructor:
  *	- `loadTest`: an object with the following attributes:
  *		- `latency`: a variable to measure latency.
  *		- `options`: same options as exports.loadTest.
  *		- `running`: if the loadTest is running or not.
  */
-export class NetworkClient {
+export class TcpClient {
 	constructor(loadTest) {
 		this.loadTest = loadTest
 		this.latency = loadTest.latency
@@ -53,7 +53,7 @@ export class NetworkClient {
 		this.params.request = this.createRequest()
 		for (const option of forbiddenOptions) {
 			if (this.options[option]) {
-				throw new Error(`${option} not supported with network sockets`)
+				throw new Error(`${option} not supported with TCP sockets`)
 			}
 		}
 	}
