+++ lib/testserver.js
@@ -60,7 +60,7 @@ class TestServer {
 	 * The callback parameter will be called after the server has started.
 	 */
 	start(callback) {
-		if (this.options.socket) {
+		if (this.options.tcp) {
 			// just for internal debugging
 			this.server = net.createServer(() => this.socketListen());
 		} else {
